Move error to module constants.
